// src/models/sortValue.ts
export enum SortValue {
  NameAsc = 'NameAsc',
  NameDesc = 'NameDesc',
  PopularityAsc = 'PopularityAsc',
  PopularityDesc = 'PopularityDesc',
  PriceAsc = 'PriceAsc',
  PriceDesc = 'PriceDesc',
  AvailableQuantityAsc = 'AvailableQuantityAsc',
  AvailableQuantityDesc = 'AvailableQuantityDesc',
  YearOfIssueAsc = 'YearOfIssueAsc',
  YearOfIssueDesc = 'YearOfIssueDesc',
}